+++ src/kimi_cli/soul/context.py
@@ -127,6 +127,33 @@ class Context:
                     message = Message.model_validate(line_json)
                     self._history.append(message)
 
+    async def clear(self):
+        """
+        Clear the context history.
+        This is almost equivalent to revert_to(0), but without relying on the assumption
+        that the first checkpoint exists.
+        File backend will be rotated.
+
+        Raises:
+            RuntimeError: When no available rotation path is found.
+        """
+
+        logger.debug("Clearing context")
+
+        # rotate the context file
+        rotated_file_path = await next_available_rotation(self._file_backend)
+        if rotated_file_path is None:
+            logger.error("No available rotation path found")
+            raise RuntimeError("No available rotation path found")
+        await aiofiles.os.replace(self._file_backend, rotated_file_path)
+        logger.debug(
+            "Rotated context file: {rotated_file_path}", rotated_file_path=rotated_file_path
+        )
+
+        self._history.clear()
+        self._token_count = 0
+        self._next_checkpoint_id = 0
+
     async def append_message(self, message: Message | Sequence[Message]):
         logger.debug("Appending message(s) to context: {message}", message=message)
         messages = message if isinstance(message, Sequence) else [message]